towards datacube.refactor - kamangir/bolt#746
