+++ blue_geo/QGIS/console/project.py
@@ -2,13 +2,75 @@ import os
 from tqdm import tqdm
 
 if not QGIS_is_live:
-    from .logger import Q_verbose, Q_log
+    from .file import Q_file_exists, Q_get_file_extension
+    from .logger import Q_verbose, Q_log, Q_log_error
     from .graphics import Q_refresh
     from .file_load import Q_load_yaml
-    from .mock import QgsProject
+
+    from .mock import (
+        QgsRasterLayer,
+        QgsProject,
+        QgsMapLayerStyle,
+        QgsVectorLayer,
+        iface,
+    )
 
 
 class ABCLI_QGIS_Project(object):
+    def add_layer(
+        self,
+        filename: str,
+        layer_name: str,
+        template_name: str = "",
+        set_nodata: bool = False,
+        nodata: int = 0,
+        refresh: bool = True,
+        verbose: bool = True,
+    ) -> bool:
+        if self.exists(layer_name, verbose=verbose):
+            return True
+
+        if Q_get_file_extension(filename) in ["geojson", "shp"]:
+            layer_ = QgsVectorLayer(filename, layer_name, "ogr")
+        elif filename.endswith(".tif"):
+            layer_ = QgsRasterLayer(filename, layer_name)
+        else:
+            Q_log_error(f"cannot load {filename}.")
+            return False
+
+        if not layer_.isValid():
+            Q_log_error(f"invalid layer: {filename}.")
+            return False
+
+        QgsProject.instance().addMapLayer(layer_)
+
+        if layer_ and layer_.dataProvider() and set_nodata:
+            layer_.dataProvider().setNoDataValue(1, nodata)
+            layer_.dataProvider().reloadData()
+            layer_.triggerRepaint()
+
+        if template_name:
+            template_layer = self.get_layer(template_name)
+            if not len(template_layer):
+                Q_log_error(f"template not found: {template_name}.")
+                return False
+
+            # https://gis.stackexchange.com/a/357206/210095
+            source_style = QgsMapLayerStyle()
+            source_style.readFromLayer(template_layer[0])
+            source_style.writeToLayer(layer_)
+            layer_.triggerRepaint()
+
+        Q_log(
+            f"{layer_name} += {template_name}",
+            icon="🎨",
+        )
+
+        if refresh:
+            Q_refresh()
+
+        return True
+
     def exists(
         self,
         layer_name: str,
@@ -58,7 +120,7 @@ class ABCLI_QGIS_Project(object):
     def metadata(self):
         filename = os.path.join(self.path, "metadata.yaml")
 
-        if not os.path.exists(filename):
+        if not Q_file_exists(filename):
             return {"error": f"{filename}: file not found."}
 
         return Q_load_yaml(filename)
