+++ blue_geo/__init__.py
@@ -9,7 +9,7 @@ ICON = "🌐"
 
 DESCRIPTION = f"{ICON} AI for a Blue Planet."
 
-VERSION = "4.973.1"
+VERSION = "4.975.1"
 
 REPO_NAME = "blue-geo"
